add-recommendations: refactor getRecommendationFilms

+++ src/main/java/ru/yandex/practicum/filmorate/storage/like/LikeStorage.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface LikeStorage {
 
-    List<Long> getAll(long filmId);
+    List<Integer> getAll(Integer filmId);
 
     void add(long userId, long filmId);
